+++ R/classdist.R
@@ -22,6 +22,8 @@
 #'  by pooling the data for all of the classes?
 #' @param log A logical: should the distances be transformed by
 #'  the natural log function?
+#' @param prefix A character string that defines the naming convention for
+#'  new distance columns. Defaults to `"classdist_"`. See Details below.
 #' @param objects Statistics are stored here once this step has
 #'  been trained by [prep.recipe()].
 #' @return An updated version of `recipe` with the new step
@@ -36,7 +38,8 @@
 #' @details `step_classdist` will create a new column for every
 #'  unique value of the `class` variable.
 #'  The resulting variables will not replace the original values
-#'  and have the prefix `classdist_`.
+#'  and by default have the prefix `classdist_`. The naming format can be
+#'  changed using the `prefix` argument.
 #'
 #' Note that, by default, the default covariance function requires
 #'  that each class should have at least as many rows as variables
@@ -48,6 +51,12 @@
 #' # in case of missing data...
 #' mean2 <- function(x) mean(x, na.rm = TRUE)
 #'
+#' # define naming convention
+#' rec <- recipe(Species ~ ., data = iris) %>%
+#'   step_classdist(all_predictors(), class = "Species",
+#'                  pool = FALSE, mean_func = mean2, prefix = "centroid_")
+#'
+#' # default naming
 #' rec <- recipe(Species ~ ., data = iris) %>%
 #'   step_classdist(all_predictors(), class = "Species",
 #'                  pool = FALSE, mean_func = mean2)
@@ -71,6 +80,7 @@ step_classdist <- function(recipe,
                            pool = FALSE,
                            log = TRUE,
                            objects = NULL,
+                           prefix = "classdist_",
                            skip = FALSE,
                            id = rand_id("classdist")) {
   if (!is.character(class) || length(class) != 1)
@@ -87,6 +97,7 @@ step_classdist <- function(recipe,
       pool = pool,
       log = log,
       objects = objects,
+      prefix = prefix,
       skip = skip,
       id = id
     )
@@ -95,7 +106,7 @@ step_classdist <- function(recipe,
 
 step_classdist_new <-
   function(terms, class, role, trained, mean_func,
-           cov_func, pool, log, objects, skip, id) {
+           cov_func, pool, log, objects, prefix, skip, id) {
     step(
       subclass = "classdist",
       terms = terms,
@@ -107,6 +118,7 @@ step_classdist_new <-
       pool = pool,
       log = log,
       objects = objects,
+      prefix = prefix,
       skip = skip,
       id = id
     )
@@ -151,6 +163,7 @@ prep.step_classdist <- function(x, training, info = NULL, ...) {
     pool = x$pool,
     log = x$log,
     objects = res,
+    prefix = x$prefix,
     skip = x$skip,
     id = x$id
   )
@@ -181,7 +194,7 @@ bake.step_classdist <- function(object, new_data, ...) {
   if (object$log)
     res <- lapply(res, log)
   res <- as_tibble(res)
-  newname <- paste0("classdist_", colnames(res))
+  newname <- paste0(object$prefix, colnames(res))
   res <- check_name(res, new_data, object, newname)
   res <- bind_cols(new_data, res)
   if (!is_tibble(res))