+++ tests/testthat/test_intercept.R
@@ -7,7 +7,8 @@ context("Intercept creation")
 
 context("step_intercept")
 
-ex_dat <- data.frame(cat = rep(c("A", "B"), each = 5), numer = 1:10)
+ex_dat <- data.frame(cat = factor(rep(c("A", "B"), each = 5)),
+                     numer = 1:10)
 
 test_that('add appropriate column with default settings', {
   rec <- recipe(~ ., data = ex_dat) %>%