+++ NEWS.md
@@ -1,8 +1,16 @@
-# recipes 0.1.9.9000
+# recipes (development version)
+
+# recipes 0.1.10
 
 ## Breaking Changes
 
 * renamed `yj_trans()` to `yj_transform()` to avoid conflicts. 
+
+## Other Changes
+
+* Added flexible naming options for new columns created by `step_depth()` and `step_classdist()` (#262).
+
+* Small changes for base R's `stringsAsFactors` change. 
 
 # recipes 0.1.9